requirements.txt
Flask
Flask-SQLAlchemy
Flask-Restless
flask-bootstrap
newsapi-python
googletrans==4.0.0rc1
